+++ core/src/main/scala/org/bitcoins/core/psbt/PSBT.scala
@@ -3,14 +3,11 @@ package org.bitcoins.core.psbt
 import org.bitcoins.core.crypto._
 import org.bitcoins.core.hd.BIP32Path
 import org.bitcoins.core.number.UInt32
-import org.bitcoins.core.policy.Policy
 import org.bitcoins.core.protocol.script._
 import org.bitcoins.core.protocol.transaction._
-import org.bitcoins.core.script.PreExecutionScriptProgram
 import org.bitcoins.core.script.crypto.HashType
 import org.bitcoins.core.script.interpreter.ScriptInterpreter
-import org.bitcoins.core.script.result.ScriptOk
-import org.bitcoins.core.util.BitcoinSLogger
+import org.bitcoins.core.util.{BitcoinSLogger, BitcoinScriptUtil}
 import org.bitcoins.core.wallet.signer.BitcoinSigner
 import org.bitcoins.core.wallet.utxo._
 import org.bitcoins.crypto._
@@ -198,7 +195,7 @@ case class PSBT(
             .isWitnessScriptPubKey(
               inputMap.redeemScriptOpt.get.redeemScript.asm)
         val notBIP143Vulnerable =
-          !out.scriptPubKey.isInstanceOf[BIP143VulnerableScriptPubKey]
+          !out.scriptPubKey.isInstanceOf[WitnessScriptPubKeyV0]
 
         if (
           (outIsWitnessScript || hasWitScript || hasWitRedeemScript) && notBIP143Vulnerable
@@ -603,122 +600,7 @@ case class PSBT(
     inputMaps.zipWithIndex.foldLeft(Try(extractTransaction)) {
       case (txT, (inputMap, index)) =>
         txT.flatMap { tx =>
-          val wUtxoOpt = inputMap.witnessUTXOOpt
-          val utxoOpt = inputMap.nonWitnessOrUnknownUTXOOpt
-
-          (wUtxoOpt, tx) match {
-            case (Some(wUtxo), wtx: WitnessTransaction) =>
-              val output = wUtxo.witnessUTXO
-              val txSigComponent = WitnessTxSigComponent(wtx,
-                                                         UInt32(index),
-                                                         output,
-                                                         Policy.standardFlags)
-              val inputResult =
-                ScriptInterpreter.run(PreExecutionScriptProgram(txSigComponent))
-              if (inputResult == ScriptOk) {
-                Success(tx)
-              } else {
-                Failure(
-                  new RuntimeException(
-                    s"Input $index was invalid: $inputResult"))
-              }
-            case (Some(_), _: NonWitnessTransaction) =>
-              Failure(new RuntimeException(
-                s"Extracted program is not witness transaction, but input $index has WitnessUTXO record"))
-            case (None, _: NonWitnessTransaction) =>
-              utxoOpt match {
-                case Some(utxo) =>
-                  val input = tx.inputs(index)
-                  val output = utxo.transactionSpent.outputs(
-                    input.previousOutput.vout.toInt)
-                  val txSigComponent = BaseTxSigComponent(tx,
-                                                          UInt32(index),
-                                                          output,
-                                                          Policy.standardFlags)
-                  val inputResult = ScriptInterpreter.run(
-                    PreExecutionScriptProgram(txSigComponent))
-                  if (inputResult == ScriptOk) {
-                    Success(tx)
-                  } else {
-                    Failure(
-                      new RuntimeException(
-                        s"Input $index was invalid: $inputResult"))
-                  }
-                case None =>
-                  logger.info(
-                    s"No UTXO record was provided for input $index, hence no validation was done for this input")
-
-                  Success(tx)
-              }
-            case (None, wtx: WitnessTransaction) =>
-              utxoOpt match {
-                case Some(utxo) =>
-                  val input = tx.inputs(index)
-                  val output = utxo.transactionSpent.outputs(
-                    input.previousOutput.vout.toInt)
-                  val txSigComponent = output.scriptPubKey match {
-                    case _: WitnessScriptPubKey =>
-                      WitnessTxSigComponent(wtx,
-                                            UInt32(index),
-                                            output,
-                                            Policy.standardFlags)
-                    case _: P2SHScriptPubKey =>
-                      (inputMap.finalizedScriptWitnessOpt,
-                       inputMap.finalizedScriptSigOpt) match {
-                        case (Some(_), _) =>
-                          WitnessTxSigComponent(wtx,
-                                                UInt32(index),
-                                                output,
-                                                Policy.standardFlags)
-                        case (None, Some(finalizedScriptSig)) =>
-                          finalizedScriptSig.scriptSig match {
-                            case p2sh: P2SHScriptSignature
-                                if p2sh.redeemScript
-                                  .isInstanceOf[WitnessScriptPubKey] =>
-                              WitnessTxSigComponent(wtx,
-                                                    UInt32(index),
-                                                    output,
-                                                    Policy.standardFlags)
-                            case _: ScriptSignature =>
-                              BaseTxSigComponent(wtx,
-                                                 UInt32(index),
-                                                 output,
-                                                 Policy.standardFlags)
-                          }
-
-                        case (None, None) =>
-                          BaseTxSigComponent(wtx,
-                                             UInt32(index),
-                                             output,
-                                             Policy.standardFlags)
-                      }
-                    case _: NonWitnessScriptPubKey =>
-                      BaseTxSigComponent(wtx,
-                                         UInt32(index),
-                                         output,
-                                         Policy.standardFlags)
-                  }
-                  val inputResult = ScriptInterpreter.run(
-                    PreExecutionScriptProgram(txSigComponent))
-                  if (inputResult == ScriptOk) {
-                    Success(tx)
-                  } else {
-                    println("============")
-                    println(output.scriptPubKey)
-                    println(
-                      (inputMap.witnessScriptOpt, inputMap.redeemScriptOpt))
-                    println("============")
-                    Failure(
-                      new RuntimeException(
-                        s"Input $index was invalid: $inputResult"))
-                  }
-                case None =>
-                  logger.info(
-                    s"No UTXO record was provided for input $index, hence no validation was done for this input")
-
-                  Success(tx)
-              }
-          }
+          BitcoinScriptUtil.verifyPSBTInputScript(tx, inputMap, index)
         }
     }
   }
