+++ core/src/main/scala/org/bitcoins/core/util/BitcoinScriptUtil.scala
@@ -12,12 +12,8 @@ import org.bitcoins.core.protocol.script.{
   EmptyScriptPubKey,
   _
 }
-import org.bitcoins.core.protocol.transaction.{
-  Transaction,
-  TransactionInput,
-  TransactionOutput,
-  WitnessTransaction
-}
+import org.bitcoins.core.protocol.transaction._
+import org.bitcoins.core.psbt.InputPSBTMap
 import org.bitcoins.core.script.constant._
 import org.bitcoins.core.script.crypto.{
   OP_CHECKMULTISIG,
@@ -30,7 +26,8 @@ import org.bitcoins.core.script.interpreter.ScriptInterpreter
 import org.bitcoins.core.script.result.{
   ScriptError,
   ScriptErrorPubKeyType,
-  ScriptErrorWitnessPubKeyType
+  ScriptErrorWitnessPubKeyType,
+  ScriptOk
 }
 import org.bitcoins.core.script.{
   ExecutionInProgressScriptProgram,
@@ -42,7 +39,7 @@ import org.bitcoins.crypto.{ECDigitalSignature, ECPublicKey}
 import scodec.bits.ByteVector
 
 import scala.annotation.tailrec
-import scala.util.Try
+import scala.util.{Failure, Success, Try}
 
 /**
   * Created by chris on 3/2/16.
@@ -670,6 +667,103 @@ trait BitcoinScriptUtil extends BitcoinSLogger {
         PreExecutionScriptProgram(txSigComponent)
     }
     ScriptInterpreter.runAllVerify(programs)
+  }
+
+  def verifyPSBTInputScript(
+      tx: Transaction,
+      inputMap: InputPSBTMap,
+      index: Int,
+      flags: Seq[ScriptFlag] = Policy.standardFlags): Try[Transaction] = {
+    val wUtxoOpt = inputMap.witnessUTXOOpt
+    val utxoOpt = inputMap.nonWitnessOrUnknownUTXOOpt
+
+    (wUtxoOpt, tx) match {
+      case (Some(wUtxo), wtx: WitnessTransaction) =>
+        val output = wUtxo.witnessUTXO
+        val txSigComponent =
+          WitnessTxSigComponent(wtx, UInt32(index), output, flags)
+        val inputResult =
+          ScriptInterpreter.run(PreExecutionScriptProgram(txSigComponent))
+
+        if (inputResult == ScriptOk) {
+          Success(tx)
+        } else {
+          Failure(
+            new RuntimeException(s"Input $index was invalid: $inputResult"))
+        }
+      case (Some(_), _: NonWitnessTransaction) =>
+        Failure(new RuntimeException(
+          s"Extracted program is not witness transaction, but input $index has WitnessUTXO record"))
+      case (None, _: NonWitnessTransaction) =>
+        utxoOpt match {
+          case Some(utxo) =>
+            val input = tx.inputs(index)
+            val output =
+              utxo.transactionSpent.outputs(input.previousOutput.vout.toInt)
+            val txSigComponent =
+              BaseTxSigComponent(tx, UInt32(index), output, flags)
+            val inputResult =
+              ScriptInterpreter.run(PreExecutionScriptProgram(txSigComponent))
+            if (inputResult == ScriptOk) {
+              Success(tx)
+            } else {
+              Failure(
+                new RuntimeException(s"Input $index was invalid: $inputResult"))
+            }
+          case None =>
+            logger.info(
+              s"No UTXO record was provided for input $index, hence no validation was done for this input")
+
+            Success(tx)
+        }
+      case (None, wtx: WitnessTransaction) =>
+        utxoOpt match {
+          case Some(utxo) =>
+            val input = tx.inputs(index)
+            val output =
+              utxo.transactionSpent.outputs(input.previousOutput.vout.toInt)
+
+            val txSigComponent = output.scriptPubKey match {
+              case _: WitnessScriptPubKey =>
+                WitnessTxSigComponent(wtx, UInt32(index), output, flags)
+              case _: P2SHScriptPubKey =>
+                (inputMap.finalizedScriptWitnessOpt,
+                 inputMap.finalizedScriptSigOpt) match {
+                  case (Some(_), _) =>
+                    WitnessTxSigComponent(wtx, UInt32(index), output, flags)
+                  case (None, Some(finalizedScriptSig)) =>
+                    finalizedScriptSig.scriptSig match {
+                      case p2sh: P2SHScriptSignature
+                          if p2sh.redeemScript
+                            .isInstanceOf[WitnessScriptPubKey] =>
+                        WitnessTxSigComponent(wtx, UInt32(index), output, flags)
+                      case _: ScriptSignature =>
+                        BaseTxSigComponent(wtx, UInt32(index), output, flags)
+                    }
+
+                  case (None, None) =>
+                    BaseTxSigComponent(wtx, UInt32(index), output, flags)
+                }
+              case _: NonWitnessScriptPubKey =>
+                BaseTxSigComponent(wtx, UInt32(index), output, flags)
+            }
+
+            val inputResult =
+              ScriptInterpreter.run(PreExecutionScriptProgram(txSigComponent))
+
+            if (inputResult == ScriptOk) {
+              Success(tx)
+            } else {
+              Failure(
+                new RuntimeException(s"Input $index was invalid: $inputResult"))
+            }
+          case None =>
+            logger.info(
+              s"No UTXO record was provided for input $index, hence no validation was done for this input")
+
+            Success(tx)
+        }
+    }
   }
 }
